Move /authentication/profile.js to /refactor_2021.
