+++ packages/openneuro-app/src/scripts/authentication/loginCheck.js
@@ -1,4 +1,7 @@
-import { getProfile, guardExpired } from './profile.js'
+import {
+  getProfile,
+  guardExpired,
+} from '../refactor_2021/authentication/profile.js'
 
 // Expects a universal cookie
 export const loginCheck = cookies => guardExpired(getProfile(cookies))
