+++ packages/openneuro-app/src/scripts/datalad/dataset/delete-page.jsx
@@ -5,7 +5,10 @@ import { useCookies } from 'react-cookie'
 import DeleteDatasetForm from '../mutations/delete-dataset-form.jsx'
 import DeleteDataset from '../mutations/delete.jsx'
 import LoggedIn from '../../authentication/logged-in.jsx'
-import { hasEditPermissions, getProfile } from '../../authentication/profile.js'
+import {
+  hasEditPermissions,
+  getProfile,
+} from '../../refactor_2021/authentication/profile.js'
 import styled from '@emotion/styled'
 
 const Container = styled.div({
