+++ packages/openneuro-app/src/scripts/datalad/fragments/freshdesk-widget.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { useCookies } from 'react-cookie'
-import { getProfile } from '../../authentication/profile'
+import { getProfile } from '../../refactor_2021/authentication/profile'
 import { config } from '../../config'
 
 const buildCustomQuery = (customText, prepopulatedFields) => {
